chore(service): move GenerateLottos other package
- Moved file and test file to service
- No logic changes, only package adjustment for better structure
