+++ src/main/kotlin/lotto/service/GenerateLottos.kt
@@ -1,4 +1,4 @@
-package lotto.service.logic
+package lotto.service
 
 import lotto.Lotto
